chore: converts SearchResults to typescript
